refactor UserService
