+++ BlazorMarkDownAppJwt/BlazorMarkDownAppJwt/Server/Entities/User.cs
@@ -1,0 +1,15 @@
+﻿namespace BlazorMarkDownAppJwt.Server.Entities
+{
+    public class User
+    {
+		public long Id { get; set; }
+
+		public string? Email { get; set; }
+
+		public string? Password { get; set; }
+
+		public string? FirstName { get; set; }
+
+		public string? LastName { get; set; }
+	}
+}
