+++ BlazorMarkDownAppJwt/BlazorMarkDownAppJwt/Server/Services/Users/IUserService.cs
@@ -1,0 +1,10 @@
+﻿using BlazorMarkDownAppJwt.Server.Entities;
+
+namespace BlazorMarkDownAppJwt.Server.Services.Users
+{
+    public interface IUserService
+    {
+        Task<User?> AuthenticateUser(string? email, string? password);
+        Task<User?> AddUser(User user);
+    }
+}
