+++ BlazorMarkDownAppJwt/BlazorMarkDownAppJwt/Server/Services/Users/UserService.cs
@@ -1,30 +1,15 @@
-﻿// UserDatabase.cs
+﻿using BlazorMarkDownAppJwt.Server.Entities;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace BlazorMarkDownAppJwt.Server
+namespace BlazorMarkDownAppJwt.Server.Services.Users
 {
-	public class User
-	{
-		public long Id { get; set; }
-		
-		public string? Email { get; set; }
-
-		public string? Password { get; set; }
-
-		public string? FirstName { get; set; }
-
-		public string? LastName { get; set; }
-	}
-	public interface IUserDatabase
-	{
-		Task<User?> AuthenticateUser(string? email, string? password);
-		Task<User?> AddUser(User user);
-	}
-	public class UserDatabase : IUserDatabase
-	{
+    public class UserService : IUserService
+    {
 		private readonly IWebHostEnvironment env;
-		public UserDatabase(IWebHostEnvironment env) => this.env = env;
+
+		private const string userPath = "Datas\\Json\\Users";
+		public UserService(IWebHostEnvironment env) => this.env = env;
 		private static string CreateHash(string password)
 		{
 			var salt = "997eff51db1544c7a3c2ddeb2053f052";
@@ -36,7 +21,7 @@ namespace BlazorMarkDownAppJwt.Server
 		{
 			if (string.IsNullOrEmpty(email) || string.IsNullOrEmpty(password))
 				return null;
-			var path = System.IO.Path.Combine(env.ContentRootPath, "Users");
+			var path = System.IO.Path.Combine(env.ContentRootPath, userPath);
 			if (!System.IO.Directory.Exists(path))
 				return null;
 			path = System.IO.Path.Combine(path, email);
@@ -44,7 +29,7 @@ namespace BlazorMarkDownAppJwt.Server
 				return null;
 			if (await System.IO.File.ReadAllTextAsync(path) != CreateHash(password))
 				return null;
-			return new User 
+			return new User
 			{
 				Email = email,
 			};
@@ -55,7 +40,7 @@ namespace BlazorMarkDownAppJwt.Server
 			{
 				if (string.IsNullOrEmpty(addedUser.Email) || string.IsNullOrEmpty(addedUser.Password))
 					return null;
-				var path = System.IO.Path.Combine(env.ContentRootPath, "Users"); // NOTE: THIS WILL CREATE THE "USERS" FOLDER IN THE PROJECT'S FOLDER!!!
+				var path = System.IO.Path.Combine(env.ContentRootPath, userPath); // NOTE: THIS WILL CREATE THE "USERS" FOLDER IN THE PROJECT'S FOLDER!!!
 				if (!System.IO.Directory.Exists(path))
 					System.IO.Directory.CreateDirectory(path); // NOTE: MAKE SURE THERE ARE CREATE/WRITE PERMISSIONS
 				path = System.IO.Path.Combine(path, addedUser.Email);
